Implementación del ajuste del error por BigDecimal

+++ src/main/java/ar/edu/itba/ss/integrators/CoupledIntegrator.java
@@ -1,7 +1,8 @@
 package ar.edu.itba.ss.integrators;
 import ar.edu.itba.ss.CoupledOscillators;
+import java.math.BigDecimal;
 
 public interface CoupledIntegrator {
-    void initialize(CoupledOscillators osc, double dt);
-    void step(CoupledOscillators osc, double t, double dt);
+    void initialize(CoupledOscillators osc, BigDecimal dt);
+    void step(CoupledOscillators osc, BigDecimal t, BigDecimal dt);
 }
